+++ micc/templates/package-base/{{cookiecutter.project_name}}/tests/test_{{cookiecutter.package_name}}.py
@@ -1,21 +1,14 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-"""Tests for `{{ cookiecutter.project_name }}` package."""
+"""Tests for {{ cookiecutter.package_name }} package."""
 
 import os
 import sys
 import pytest
 
-from click.testing import CliRunner
-from click import echo
-
 import {{ cookiecutter.package_name }}
-{#
-{%- if cookiecutter.command_line_interface|lower == 'click' %}
-from {{ cookiecutter.project_name }} import cli
-{%- endif %}
-#}
+
 # Make sure that the current directory is the project directory.
 # 'make test" and 'pytest' are generally run from the project directory.
 # However, if we run/debug this file in eclipse, we end up in test
@@ -31,43 +24,24 @@ if not ('.' in sys.path or os.getcwd() in sys.path):
     echo(f"Adding '.' to sys.path.\n")
     sys.path.insert(0, '.')
 
-# ==============================================================================
-@pytest.fixture
-def response():
-    """Sample pytest fixture.
+def test_hello():
+    """Test for {{ cookiecutter.package_name }}.hello()."""
+    s = {{ cookiecutter.package_name }}.hello()
+    assert s=="Hello world"
 
-    See more at: http://doc.pytest.org/en/latest/fixture.html
-    """
-    # import requests
-    # return requests.get('https://github.com/audreyr/cookiecutter-pypackage')
-# ==============================================================================
-def test_content(response):
-    """Sample pytest test function with the pytest fixture as an argument."""
-    # from bs4 import BeautifulSoup
-    # assert 'GitHub' in BeautifulSoup(response.content).title.string
-# ==============================================================================
-def test_version():
-    assert {{ cookiecutter.package_name }}.__version__=="0.0.0"
-# ==============================================================================
-# def test_command_line_interface():
-#     """Test the CLI."""
-#     runner = CliRunner()
-#     result = runner.invoke(cli.main)
-#     assert result.exit_code == 0
-#     assert '{{ cookiecutter.project_name }}.cli.main' in result.output
-#     help_result = runner.invoke(cli.main, ['--help'])
-#     assert help_result.exit_code == 0
-#     assert '--help  Show this message and exit.' in help_result.output
-
+    s = {{ cookiecutter.package_name }}.hello('me')
+    assert s=="Hello me"
+    
+    
 # ==============================================================================
 # The code below is for debugging a particular test in eclipse/pydev.
 # (normally all tests are run with pytest)
 # ==============================================================================
 if __name__ == "__main__":
-    the_test_you_want_to_debug = test_version
+    the_test_you_want_to_debug = test_hello
 
-    from execution_trace import trace
-    with trace(f"__main__ running {the_test_you_want_to_debug}",
-               '-*# finished #*-', singleline=False, combine=False):
-        the_test_you_want_to_debug()
-# ==============================================================================
+    print("__main__ running", the_test_you_want_to_debug)
+    the_test_you_want_to_debug()
+    print('-*# finished #*-')
+    
+# eof